Improve style for subscription page; Add Omnes font; Enhance structure with images and fonts folders; Rename some folders
