+++ frontend/src/App.css
@@ -5,16 +5,16 @@
   margin: 0;
   padding: 0;
   box-sizing: border-box;
+  font-family: var(--font-primary);
 }
 
 button {
-  padding: 0.75rem 1.5rem;
+  padding: 0.75rem 1.25rem;
   font-size: 1rem;
   background-color: var(--color-ocean);
   color: white;
   border: none;
   border-radius: 8px;
-  cursor: pointer;
   transition: background-color 0.3s ease;
 }
 
@@ -22,7 +22,6 @@ button:hover {
   background-color: var(--color-sky);
 }
 
-/* ovo je globalno za sve stranice s formama */
 .page-container {
   background-color: var(--color-cloud);
   font-family: var(--font-primary);
