+++ frontend/src/assets/images/Minus icon.png
