+++ frontend/src/assets/images/lighthouse.png
