+++ frontend/src/assets/images/ridge.png
