+++ frontend/src/components/Map/buttons.ts
@@ -1,9 +1,9 @@
-import sosButton from "../../assets/Sos icon.svg";
-import plusButton from "../../assets/Plus icon.svg";
-import minusButton from "../../assets/Minus icon.svg";
-import compassButton from "../../assets/Navigation icon.svg";
-import profileButton from "../../assets/Profile icon.svg";
-import infoButton from "../../assets/Rules and help icon.svg";
+import sosButton from "../../assets/images/Sos icon.svg";
+import plusButton from "../../assets/images/Plus icon.svg";
+import minusButton from "../../assets/images/Minus icon.svg";
+import compassButton from "../../assets/images/Navigation icon.svg";
+import profileButton from "../../assets/images/Profile icon.svg";
+import infoButton from "../../assets/images/Rules and help icon.svg";
 
 export {
   sosButton,
