+++ frontend/src/components/PlanButton/PlanButton.module.css
@@ -1,42 +1,85 @@
 .planButton {
   all: unset;
+  margin: 0 auto;
   display: flex;
   flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  width: 100%;
-  padding: 1.5rem 1rem;
-  border: 2px solid #e0e0e0;
-  border-radius: 8px;
-  background-color: white;
-  color: #333;
-  cursor: pointer;
-  transition: all 0.3s ease;
-  position: relative;
+  justify-content: space-between;
+  width: 90%;
+  padding: 1.25rem 1rem;
+  border: 1.5px solid #ddd;
+  border-radius: 10px;
   box-sizing: border-box;
-  margin-bottom: 1rem;
-  text-align: center;
+  text-align: left;
+  position: relative;
+  transition: border 0.3s;
+}
+
+.planButton.selected {
+  border-color: #000;
+}
+
+.planButton:hover {
+  background-color: white;
+}
+
+.topRow {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  margin-bottom: 0.5rem;
+}
+
+.period,
+.price {
+  color: #111;
+  font-size: 16px;
+  font-weight: 500;
+  line-height: 24px;
 }
 
 .checkCircle {
-  position: absolute;
-  top: 10px;
-  right: 10px;
-  width: 24px;
-  height: 24px;
   display: flex;
-  justify-content: center;
   align-items: center;
+  justify-content: center;
 }
 
-.priceLabel {
-  font-size: 0.875rem;
-  color: #666;
-  margin-bottom: 0.25rem;
+.circle {
+  width: 20px;
+  height: 20px;
+  border-radius: 50%;
+  background-color: white;
+  border: 2px solid #ccc;
+  display: flex;
+  align-items: center;
+  justify-content: center;
 }
 
-.priceValue {
-  font-size: 1.5rem;
-  font-weight: 700;
-  color: #333;
+.circleSelected {
+  background-color: #000;
+  border-color: #000;
+}
+
+.checkIcon {
+  width: 14px;
+  height: 14px;
+}
+
+.bottomRow {
+  display: flex;
+  justify-content: space-between;
+  align-items: flex-end;
+}
+
+.subPrice {
+  display: flex;
+  flex-direction: column;
+  align-items: flex-end;
+  font-size: 0.9rem;
+  color: #555;
+  line-height: 1.2;
+}
+
+.subPriceText {
+  font-size: 0.8rem;
+  color: #999;
 }
