+++ frontend/src/pages/BoatPages/InfoBoatPage.tsx
