+++ frontend/src/pages/RegisterPages/SubscriptionPage.module.css
@@ -6,52 +6,24 @@
 }
 
 .headline {
-  margin-top: 20px;
+  margin-top: 50px;
+  color: var(--color-ocean);
   font-size: 2rem;
   margin-bottom: 0.5rem;
-  font-weight: bold;
+  font-weight: 600;
 }
 
 .trialText {
   margin-top: 50px;
   font-size: 1.2rem;
   margin-bottom: 2rem;
-  font-weight: 600;
 }
 
 .plansContainer {
   display: flex;
   flex-direction: column;
   gap: 1rem;
+  margin: 0 auto;
   margin-bottom: 2rem;
   max-width: 500px;
-  margin-left: auto;
-  margin-right: auto;
-}
-
-.nextButton {
-  background-color: #1976d2;
-  color: white;
-  border: none;
-  border-radius: 4px;
-  padding: 0.75rem 2rem;
-  font-size: 1rem;
-  font-weight: 600;
-  cursor: pointer;
-  transition: background-color 0.3s;
-}
-
-.nextButton:hover {
-  background-color: #1565c0;
-}
-
-.nextButton:disabled {
-  background-color: #ccc;
-  cursor: not-allowed;
-}
-
-@media (max-width: 600px) {
-  .plansContainer {
-    flex-direction: column;
-  }
 }
